split up WebsocketConnection.tsx to pass lint rule
